IOS-4943 Remove SettingsFactory
